// src/app/azure-storage/types/azure-storage.d.ts
import { BlobServiceClient } from '@azure/storage-blob';

export interface BlobItem {
  name: string;
  containerName: string;
}

export interface BlobItemDownload extends BlobItem {
  url: string;
}

export interface BlobItemUpload extends BlobItem {
  progress: number;
}

export interface BlobStorageRequest {
  storageUri: string;
  storageAccessToken: string;
}

export interface BlobContainerRequest extends BlobStorageRequest {
  containerName: string;
}

export interface BlobFileRequest extends BlobContainerRequest {
  filename: string;
}

export type Dictionary<T> = { [key: string]: T };

export type BlobStorageClientFactory = (
  options: BlobStorageRequest
) => BlobServiceClient;